fix: eliminate architecture violation in cli_utils (#58)
Moved check_and_auto_sync() from core/cli_utils.py to entrypoints/cli.py
to fix circular dependency and restore clean architecture.

Changes:
- Move check_and_auto_sync() to entrypoints/cli.py (presentation layer)
- Remove imports from adapters/ and entrypoints/ in core/cli_utils.py
- Simplify nested conditionals in moved function (ruff SIM102)
- core/ layer now only depends on ports/, never adapters/

Architecture fix: core/ depends ONLY on ports/, never adapters/
No user-facing changes - internal refactoring only.
All tests pass.
